fixed bugs

+++ schemas/DishProduct.go
@@ -1,8 +1,8 @@
 package schemas
 
 type DishProduct struct {
-	ProductId         int    `json:"product-id"`
-	Name              string `json:"name"`
-	Amount            int    `json:"amount"`
-	UnitOfMeasurement string `json:"unit-of-measurement"`
+	ProductId         int     `json:"product-id"`
+	Name              string  `json:"name"`
+	Amount            float32 `json:"amount"`
+	UnitOfMeasurement string  `json:"unit-of-measurement"`
 }
